requirements.txt
streamlit
tf-keras
deepface
pillow
numpy
opencv-python